Test for augmented fastq reader
