+++ kevlar/__init__.py
@@ -8,6 +8,7 @@
 # -----------------------------------------------------------------------------
 
 from . import fasta
+from .fasta import parse_augmented_fastq
 from . import dump
 from . import find
 from . import collect
@@ -39,27 +40,3 @@ def revcommin(seq):
     rc = revcom(seq)
     minseq = sorted((seq, rc))[0]
     return minseq
-
-
-def parse_augmented_fastq(instream):
-    record = None
-    annot_kmers = dict()
-
-    for line in instream:
-        if line.startswith('@'):
-            if record is not None:
-                yield record, annot_kmers
-                annot_kmers = dict()
-            readid = line[1:].strip()
-            seq = next(instream).strip()
-            _ = next(instream)
-            qual = next(instream).strip()
-            record = screed.Record(name=readid, sequence=seq, quality=qual)
-        elif line.endswith('#\n'):
-            offset = len(line) - len(line.lstrip())
-            line = line.strip()[:-1]
-            abundances = re.split('\s+', line)
-            kmer = abundances.pop(0)
-            annot_kmers[offset] = (kmer, abundances)
-    if record is not None:
-        yield record, annot_kmers
