+++ kevlar/fasta.py
@@ -8,6 +8,8 @@
 # -----------------------------------------------------------------------------
 
 from __future__ import print_function
+import re
+import screed
 
 
 def parse(data):
@@ -39,3 +41,27 @@ def parse_seq_dict(data):
         assert seqid not in seqs
         seqs[seqid] = sequence
     return seqs
+
+
+def parse_augmented_fastq(instream):
+    record = None
+    annot_kmers = dict()
+
+    for line in instream:
+        if line.startswith('@'):
+            if record is not None:
+                yield record, annot_kmers
+                annot_kmers = dict()
+            readid = line[1:].strip()
+            seq = next(instream).strip()
+            _ = next(instream)
+            qual = next(instream).strip()
+            record = screed.Record(name=readid, sequence=seq, quality=qual)
+        elif line.endswith('#\n'):
+            offset = len(line) - len(line.lstrip())
+            line = line.strip()[:-1]
+            abundances = re.split('\s+', line)
+            kmer = abundances.pop(0)
+            annot_kmers[offset] = (kmer, abundances)
+    if record is not None:
+        yield record, annot_kmers
